fix: Correctly convert parameter types
A weird thing was happening with the --batch flag.  Because the batch
size was being passed in as string, when `start + size` was executed
inside of the `batch()` function, string concatenation was being
performed instead of addition.

Converting the parameters to the proper types before passing them
to the `generateNames()` function fixes that problem.